requirements.txt
google_api_python_client==2.118.0
google_auth_oauthlib==1.2.0
protobuf==4.25.3
Requests==2.31.0
